requirements.txt
requests
python-dotenv
beautifulsoup4
scikit-learn
urllib3